Remove CoroutineState->nestingLevel
In certain scenarios yielding more than 3 Promises in an Amp context
would result in the 4th and consecutive Promises to never resolve. This
appears to have been due to an arbitrary nestingLevel check that
determined when Amp\__coroutineSend was called. For this scenario
Amp\__coroutineNextTick should be invoked as the Generator has not
finished yielding.

This patch changes the nestingLevel check to now run Amp\__coroutineSend
based on Generator->valid, ensuring that all values are yielded from the
Generator. As the CoroutineState->nestingLevel value was only used for
these checks it was removed from the CoroutineState.